Enum webrender_api::api::ApiMsg

pub enum ApiMsg {
    UpdateResources(Vec<ResourceUpdate>),
    GetGlyphDimensions(FontInstanceKey, Vec<GlyphIndex>, MsgSender<Vec<Option<GlyphDimensions>>>),
    GetGlyphIndices(FontKey, String, MsgSender<Vec<Option<u32>>>),
    CloneApi(MsgSender<IdNamespace>),
    CloneApiByClient(IdNamespace),
    AddDocument(DocumentId, DeviceIntSize, DocumentLayer),
    UpdateDocuments(Vec<DocumentId>, Vec<TransactionMsg>),
    DeleteDocument(DocumentId),
    ExternalEvent(ExternalEvent),
    ClearNamespace(IdNamespace),
    MemoryPressure,
    ReportMemory(MsgSender<MemoryReport>),
    DebugCommand(DebugCommand),
    WakeUp,
    WakeSceneBuilder,
    FlushSceneBuilder(MsgSender<()>),
    ShutDown(Option<MsgSender<()>>),
}

Variants

UpdateResources(Vec<ResourceUpdate>)

Add/remove/update images and fonts.

GetGlyphDimensions(FontInstanceKey, Vec<GlyphIndex>, MsgSender<Vec<Option<GlyphDimensions>>>)

Gets the glyph dimensions

GetGlyphIndices(FontKey, String, MsgSender<Vec<Option<u32>>>)

Gets the glyph indices from a string

CloneApi(MsgSender<IdNamespace>)

Adds a new document namespace.

CloneApiByClient(IdNamespace)

Adds a new document namespace.

AddDocument(DocumentId, DeviceIntSize, DocumentLayer)

Adds a new document with given initial size.

UpdateDocuments(Vec<DocumentId>, Vec<TransactionMsg>)

A message targeted at a particular document.

DeleteDocument(DocumentId)

Deletes an existing document.

ExternalEvent(ExternalEvent)

An opaque handle that must be passed to the render notifier. It is used by Gecko to forward gecko-specific messages to the render thread preserving the ordering within the other messages.

ClearNamespace(IdNamespace)

Removes all resources associated with a namespace.

MemoryPressure

Flush from the caches anything that isn't necessary, to free some memory.

ReportMemory(MsgSender<MemoryReport>)

Collects a memory report.

DebugCommand(DebugCommand)

Change debugging options.

WakeUp

Wakes the render backend's event loop up. Needed when an event is communicated through another channel.

WakeSceneBuilder

FlushSceneBuilder(MsgSender<()>)

ShutDown(Option<MsgSender<()>>)