Enum webrender_api::api::DebugCommand

pub enum DebugCommand {

    SetFlags(DebugFlags),
    EnableDualSourceBlending(bool),
    FetchDocuments,
    FetchPasses,
    FetchClipScrollTree,
    FetchRenderTasks,
    FetchScreenshot,
    SaveCapture(PathBuf, CaptureBits),
    LoadCapture(PathBuf, Option<(u32, u32)>, Sender<CapturedDocument>),
    StartCaptureSequence(PathBuf, CaptureBits),
    StopCaptureSequence,
    ClearCaches(ClearCache),
    EnableNativeCompositor(bool),
    EnableMultithreading(bool),
    SetBatchingLookback(u32),
    InvalidateGpuCache,
    SimulateLongSceneBuild(u32),
    SimulateLongLowPrioritySceneBuild(u32),
    SetPictureTileSize(Option<DeviceIntSize>),

}

Update of the state of built-in debugging facilities.

Variants

SetFlags

Sets the provided debug flags.

Tuple Fields of SetFlags

0: DebugFlags

EnableDualSourceBlending

Configure if dual-source blending is used, if available.

Tuple Fields of EnableDualSourceBlending

0: bool

FetchDocuments

Fetch current documents and display lists.

FetchPasses

Fetch current passes and batches.

FetchClipScrollTree

Fetch the spatial tree.

FetchRenderTasks

Fetch render tasks.

FetchScreenshot

Fetch screenshot.

SaveCapture

Save a capture of all the documents state.

Tuple Fields of SaveCapture

0: PathBuf1: CaptureBits

LoadCapture

Load a capture of all the documents state.

Tuple Fields of LoadCapture

0: PathBuf1: Option<(u32, u32)>2: Sender<CapturedDocument>

StartCaptureSequence

Start capturing a sequence of scene/frame changes.

Tuple Fields of StartCaptureSequence

0: PathBuf1: CaptureBits

StopCaptureSequence

Stop capturing a sequence of scene/frame changes.

ClearCaches

Clear cached resources, forcing them to be re-uploaded from templates.

Tuple Fields of ClearCaches

0: ClearCache

EnableNativeCompositor

Enable/disable native compositor usage

Tuple Fields of EnableNativeCompositor

0: bool

EnableMultithreading

Enable/disable parallel job execution with rayon.

Tuple Fields of EnableMultithreading

0: bool

SetBatchingLookback

Sets the maximum amount of existing batches to visit before creating a new one.

Tuple Fields of SetBatchingLookback

0: u32

InvalidateGpuCache

Invalidate GPU cache, forcing the update from the CPU mirror.

SimulateLongSceneBuild

Causes the scene builder to pause for a given amount of milliseconds each time it processes a transaction.

Tuple Fields of SimulateLongSceneBuild

0: u32

SimulateLongLowPrioritySceneBuild

Causes the low priority scene builder to pause for a given amount of milliseconds each time it processes a transaction.

Tuple Fields of SimulateLongLowPrioritySceneBuild

0: u32

SetPictureTileSize

Set an override tile size to use for picture caches

Tuple Fields of SetPictureTileSize

0: Option<DeviceIntSize>