Struct webrender_api::api::RenderApi

pub struct RenderApi {
    api_sender: Sender<ApiMsg>,
    namespace_id: IdNamespace,
    next_id: Cell<ResourceId>,
    resources: ApiResources,
}

The main entry point to interact with WebRender.

Fields
api_sender: Sender<ApiMsg>namespace_id: IdNamespacenext_id: Cell<ResourceId>resources: ApiResources
Implementations

impl RenderApi

pub fn get_namespace_id(&self) -> IdNamespace

Returns the namespace ID used by this API object.

pub fn create_sender(&self) -> RenderApiSender

pub fn add_document(
&self,
initial_size: DeviceIntSize,
layer: DocumentLayer
) -> DocumentId

Add a document to the WebRender instance.

Instances can manage one or several documents (using the same render backend thread). Each document will internally correspond to a single scene, and scenes are made of one or several pipelines.

pub fn add_document_with_id(
&self,
initial_size: DeviceIntSize,
layer: DocumentLayer,
id: u32
) -> DocumentId

See add_document

pub fn delete_document(&self, document_id: DocumentId)

Delete a document.

pub fn generate_font_key(&self) -> FontKey

Generate a new font key

pub fn generate_font_instance_key(&self) -> FontInstanceKey

Generate a new font instance key

pub fn get_glyph_dimensions(
&self,
key: FontInstanceKey,
glyph_indices: Vec<GlyphIndex>
) -> Vec<Option<GlyphDimensions>>

Gets the dimensions for the supplied glyph keys

Note: Internally, the internal texture cache doesn’t store ‘empty’ textures (height or width = 0) This means that glyph dimensions e.g. for spaces (’ ’) will mostly be None.

pub fn get_glyph_indices(&self, key: FontKey, text: &str) -> Vec<Option<u32>>

Gets the glyph indices for the supplied string. These can be used to construct GlyphKeys.

pub fn generate_image_key(&self) -> ImageKey

Creates an ImageKey.

pub fn generate_blob_image_key(&self) -> BlobImageKey

Creates a BlobImageKey.

pub fn send_external_event(&self, evt: ExternalEvent)

A Gecko-specific notification mechanism to get some code executed on the Renderer’s thread, mostly replaced by NotificationHandler. You should probably use the latter instead.

pub fn notify_memory_pressure(&self)

Notify WebRender that now is a good time to flush caches and release as much memory as possible.

pub fn report_memory(&self) -> MemoryReport

Synchronously requests memory report.

pub fn set_debug_flags(&self, flags: DebugFlags)

Update debugging flags.

pub fn shut_down(&self, synchronously: bool)

Shut the WebRender instance down.

pub fn generate_property_binding_key<T: Copy>(&self) -> PropertyBindingKey<T>

Create a new unique key that can be used for animated property bindings.

fn next_unique_id(&self) -> u32

fn frame_message(
&self,
msg: FrameMsg,
document_id: DocumentId
) -> Box<TransactionMsg>

Creates a transaction message from a single frame message.

fn scene_message(
&self,
msg: SceneMsg,
document_id: DocumentId
) -> Box<TransactionMsg>

Creates a transaction message from a single scene message.

fn send_scene_msg(&self, document_id: DocumentId, msg: SceneMsg)

A helper method to send document messages.

fn send_frame_msg(&self, document_id: DocumentId, msg: FrameMsg)

A helper method to send document messages.

pub fn send_transaction(
&mut self,
document_id: DocumentId,
transaction: Transaction
)

Send a transaction to WebRender.

pub fn send_transactions(
&mut self,
document_ids: Vec<DocumentId>,
transactions: Vec<Transaction>
)

Send multiple transactions.

pub fn hit_test(
&self,
document_id: DocumentId,
pipeline_id: Option<PipelineId>,
point: WorldPoint,
flags: HitTestFlags
) -> HitTestResult

Does a hit test on display items in the specified document, at the given point. If a pipeline_id is specified, it is used to further restrict the hit results so that only items inside that pipeline are matched. If the HitTestFlags argument contains the FIND_ALL flag, then the vector of hit results will contain all display items that match, ordered from front to back.

pub fn request_hit_tester(&self, document_id: DocumentId) -> HitTesterRequest

Synchronously request an object that can perform fast hit testing queries.

pub fn set_document_view(
&self,
document_id: DocumentId,
device_rect: DeviceIntRect,
device_pixel_ratio: f32
)

Setup the output region in the framebuffer for a given document.

pub fn enable_frame_output(
&self,
document_id: DocumentId,
pipeline_id: PipelineId,
enable: bool
)

Setup the output region in the framebuffer for a given document. Enable copying of the output of this pipeline id to an external texture for callers to consume.

pub fn get_scroll_node_state(
&self,
document_id: DocumentId
) -> Vec<ScrollNodeState>

pub fn flush_scene_builder(&self)

Block until a round-trip to the scene builder thread has completed. This ensures that any transactions (including ones deferred to the scene builder thread) have been processed.

pub fn save_capture(&self, path: PathBuf, bits: CaptureBits)

Save a capture of the current frame state for debugging.

pub fn load_capture(
&self,
path: PathBuf,
ids: Option<(u32, u32)>
) -> Vec<CapturedDocument>

Load a capture of the current frame state for debugging.

pub fn start_capture_sequence(&self, path: PathBuf, bits: CaptureBits)

Start capturing a sequence of frames.

pub fn stop_capture_sequence(&self)

Stop capturing sequences of frames.

pub fn send_debug_cmd(&mut self, cmd: DebugCommand)

Update the state of builtin debugging facilities.