webrender_api

Module channel

Re-exports

pub use crossbeam_channel as crossbeam;

Structs

MsgReceiver

MsgSender

Payload

Receiver

The receiving side of a channel.

Sender

The sending side of a channel.

Functions

fast_channel

A fast MPMC message channel that can hold a fixed number of messages.

msg_channel

payload_channel

single_msg_channel

A create a channel intended for one-shot uses, for example the channels created to block on a synchronous query and then discarded,

unbounded_channel

Creates an MPMC channel that is a bit slower than the fast_channel but doesn’t have a limit on the number of messages held at a given time and therefore doesn’t block when sending. Creates a channel of unbounded capacity.

Type Aliases

PayloadReceiver

PayloadSender