Enum webrender_api::display_item::FilterOp

#[repr(C)]
pub enum FilterOp {

    Identity,
    Blur(f32, f32),
    Brightness(f32),
    Contrast(f32),
    Grayscale(f32),
    HueRotate(f32),
    Invert(f32),
    Opacity(PropertyBinding<f32>, f32),
    Saturate(f32),
    Sepia(f32),
    DropShadow(Shadow),
    ColorMatrix([f32; 20]),
    SrgbToLinear,
    LinearToSrgb,
    ComponentTransfer,
    Flood(ColorF),
    SVGFESourceGraphic {
        node: FilterOpGraphNode,
    },
    SVGFESourceAlpha {
        node: FilterOpGraphNode,
    },
    SVGFEIdentity {
        node: FilterOpGraphNode,
    },
    SVGFEOpacity {
        node: FilterOpGraphNode,
        valuebinding: PropertyBinding<f32>,
        value: f32,
    },
    SVGFEToAlpha {
        node: FilterOpGraphNode,
    },
    SVGFEBlendDarken {
        node: FilterOpGraphNode,
    },
    SVGFEBlendLighten {
        node: FilterOpGraphNode,
    },
    SVGFEBlendMultiply {
        node: FilterOpGraphNode,
    },
    SVGFEBlendNormal {
        node: FilterOpGraphNode,
    },
    SVGFEBlendScreen {
        node: FilterOpGraphNode,
    },
    SVGFEBlendOverlay {
        node: FilterOpGraphNode,
    },
    SVGFEBlendColorDodge {
        node: FilterOpGraphNode,
    },
    SVGFEBlendColorBurn {
        node: FilterOpGraphNode,
    },
    SVGFEBlendHardLight {
        node: FilterOpGraphNode,
    },
    SVGFEBlendSoftLight {
        node: FilterOpGraphNode,
    },
    SVGFEBlendDifference {
        node: FilterOpGraphNode,
    },
    SVGFEBlendExclusion {
        node: FilterOpGraphNode,
    },
    SVGFEBlendHue {
        node: FilterOpGraphNode,
    },
    SVGFEBlendSaturation {
        node: FilterOpGraphNode,
    },
    SVGFEBlendColor {
        node: FilterOpGraphNode,
    },
    SVGFEBlendLuminosity {
        node: FilterOpGraphNode,
    },
    SVGFEColorMatrix {
        node: FilterOpGraphNode,
        values: [f32; 20],
    },
    SVGFEComponentTransfer {
        node: FilterOpGraphNode,
    },
    SVGFECompositeArithmetic {
        node: FilterOpGraphNode,
        k1: f32,
        k2: f32,
        k3: f32,
        k4: f32,
    },
    SVGFECompositeATop {
        node: FilterOpGraphNode,
    },
    SVGFECompositeIn {
        node: FilterOpGraphNode,
    },
    SVGFECompositeLighter {
        node: FilterOpGraphNode,
    },
    SVGFECompositeOut {
        node: FilterOpGraphNode,
    },
    SVGFECompositeOver {
        node: FilterOpGraphNode,
    },
    SVGFECompositeXOR {
        node: FilterOpGraphNode,
    },
    SVGFEConvolveMatrixEdgeModeDuplicate {
        node: FilterOpGraphNode,
        order_x: i32,
        order_y: i32,
        kernel: [f32; 25],
        divisor: f32,
        bias: f32,
        target_x: i32,
        target_y: i32,
        kernel_unit_length_x: f32,
        kernel_unit_length_y: f32,
        preserve_alpha: i32,
    },
    SVGFEConvolveMatrixEdgeModeNone {
        node: FilterOpGraphNode,
        order_x: i32,
        order_y: i32,
        kernel: [f32; 25],
        divisor: f32,
        bias: f32,
        target_x: i32,
        target_y: i32,
        kernel_unit_length_x: f32,
        kernel_unit_length_y: f32,
        preserve_alpha: i32,
    },
    SVGFEConvolveMatrixEdgeModeWrap {
        node: FilterOpGraphNode,
        order_x: i32,
        order_y: i32,
        kernel: [f32; 25],
        divisor: f32,
        bias: f32,
        target_x: i32,
        target_y: i32,
        kernel_unit_length_x: f32,
        kernel_unit_length_y: f32,
        preserve_alpha: i32,
    },
    SVGFEDiffuseLightingDistant {
        node: FilterOpGraphNode,
        surface_scale: f32,
        diffuse_constant: f32,
        kernel_unit_length_x: f32,
        kernel_unit_length_y: f32,
        azimuth: f32,
        elevation: f32,
    },
    SVGFEDiffuseLightingPoint {
        node: FilterOpGraphNode,
        surface_scale: f32,
        diffuse_constant: f32,
        kernel_unit_length_x: f32,
        kernel_unit_length_y: f32,
        x: f32,
        y: f32,
        z: f32,
    },
    SVGFEDiffuseLightingSpot {

        node: FilterOpGraphNode,
        surface_scale: f32,
        diffuse_constant: f32,
        kernel_unit_length_x: f32,
        kernel_unit_length_y: f32,
        x: f32,
        y: f32,
        z: f32,
        points_at_x: f32,
        points_at_y: f32,
        points_at_z: f32,
        cone_exponent: f32,
        limiting_cone_angle: f32,
    
},
    SVGFEDisplacementMap {
        node: FilterOpGraphNode,
        scale: f32,
        x_channel_selector: u32,
        y_channel_selector: u32,
    },
    SVGFEDropShadow {
        node: FilterOpGraphNode,
        color: ColorF,
        dx: f32,
        dy: f32,
        std_deviation_x: f32,
        std_deviation_y: f32,
    },
    SVGFEFlood {
        node: FilterOpGraphNode,
        color: ColorF,
    },
    SVGFEGaussianBlur {
        node: FilterOpGraphNode,
        std_deviation_x: f32,
        std_deviation_y: f32,
    },
    SVGFEImage {
        node: FilterOpGraphNode,
        sampling_filter: u32,
        matrix: [f32; 6],
    },
    SVGFEMorphologyDilate {
        node: FilterOpGraphNode,
        radius_x: f32,
        radius_y: f32,
    },
    SVGFEMorphologyErode {
        node: FilterOpGraphNode,
        radius_x: f32,
        radius_y: f32,
    },
    SVGFEOffset {
        node: FilterOpGraphNode,
        offset_x: f32,
        offset_y: f32,
    },
    SVGFESpecularLightingDistant {
        node: FilterOpGraphNode,
        surface_scale: f32,
        specular_constant: f32,
        specular_exponent: f32,
        kernel_unit_length_x: f32,
        kernel_unit_length_y: f32,
        azimuth: f32,
        elevation: f32,
    },
    SVGFESpecularLightingPoint {
        node: FilterOpGraphNode,
        surface_scale: f32,
        specular_constant: f32,
        specular_exponent: f32,
        kernel_unit_length_x: f32,
        kernel_unit_length_y: f32,
        x: f32,
        y: f32,
        z: f32,
    },
    SVGFESpecularLightingSpot {

        node: FilterOpGraphNode,
        surface_scale: f32,
        specular_constant: f32,
        specular_exponent: f32,
        kernel_unit_length_x: f32,
        kernel_unit_length_y: f32,
        x: f32,
        y: f32,
        z: f32,
        points_at_x: f32,
        points_at_y: f32,
        points_at_z: f32,
        cone_exponent: f32,
        limiting_cone_angle: f32,
    
},
    SVGFETile {
        node: FilterOpGraphNode,
    },
    SVGFETurbulenceWithFractalNoiseWithNoStitching {
        node: FilterOpGraphNode,
        base_frequency_x: f32,
        base_frequency_y: f32,
        num_octaves: u32,
        seed: u32,
    },
    SVGFETurbulenceWithFractalNoiseWithStitching {
        node: FilterOpGraphNode,
        base_frequency_x: f32,
        base_frequency_y: f32,
        num_octaves: u32,
        seed: u32,
    },
    SVGFETurbulenceWithTurbulenceNoiseWithNoStitching {
        node: FilterOpGraphNode,
        base_frequency_x: f32,
        base_frequency_y: f32,
        num_octaves: u32,
        seed: u32,
    },
    SVGFETurbulenceWithTurbulenceNoiseWithStitching {
        node: FilterOpGraphNode,
        base_frequency_x: f32,
        base_frequency_y: f32,
        num_octaves: u32,
        seed: u32,
    },

}

Variants

Identity

Filter that does no transformation of the colors, needed for debug purposes, and is the default value in impl_default_for_enums. parameters: none CSS filter semantics - operates on previous picture, uses sRGB space (non-linear)

Blur(f32, f32)

apply blur effect parameters: stdDeviationX, stdDeviationY CSS filter semantics - operates on previous picture, uses sRGB space (non-linear)

Brightness(f32)

apply brightness effect parameters: amount CSS filter semantics - operates on previous picture, uses sRGB space (non-linear)

Contrast(f32)

apply contrast effect parameters: amount CSS filter semantics - operates on previous picture, uses sRGB space (non-linear)

Grayscale(f32)

fade image toward greyscale version of image parameters: amount CSS filter semantics - operates on previous picture, uses sRGB space (non-linear)

HueRotate(f32)

fade image toward hue-rotated version of image (rotate RGB around color wheel) parameters: angle CSS filter semantics - operates on previous picture, uses sRGB space (non-linear)

Invert(f32)

fade image toward inverted image (1 - RGB) parameters: amount CSS filter semantics - operates on previous picture, uses sRGB space (non-linear)

Opacity(PropertyBinding<f32>, f32)

multiplies color and alpha by opacity parameters: amount CSS filter semantics - operates on previous picture, uses sRGB space (non-linear)

Saturate(f32)

multiply saturation of colors parameters: amount CSS filter semantics - operates on previous picture, uses sRGB space (non-linear)

Sepia(f32)

fade image toward sepia tone version of image parameters: amount CSS filter semantics - operates on previous picture, uses sRGB space (non-linear)

DropShadow(Shadow)

add drop shadow version of image to the image parameters: shadow CSS filter semantics - operates on previous picture, uses sRGB space (non-linear)

ColorMatrix([f32; 20])

transform color and alpha in image through 4x5 color matrix (transposed for efficiency) parameters: matrix[5][4] CSS filter semantics - operates on previous picture, uses sRGB space (non-linear)

SrgbToLinear

internal use - convert sRGB input to linear output parameters: none CSS filter semantics - operates on previous picture, uses sRGB space (non-linear)

LinearToSrgb

internal use - convert linear input to sRGB output parameters: none CSS filter semantics - operates on previous picture, uses sRGB space (non-linear)

ComponentTransfer

remap RGBA with color gradients and component swizzle parameters: FilterData CSS filter semantics - operates on previous picture, uses sRGB space (non-linear)

Flood(ColorF)

replace image with a solid color NOTE: UNUSED; Gecko never produces this filter parameters: color CSS filter semantics - operates on previous picture,uses sRGB space (non-linear)

SVGFESourceGraphic

Filter that copies the SourceGraphic image into the specified subregion, This is intentionally the only way to get SourceGraphic into the graph, as the filter region must be applied before it is used. parameters: FilterOpGraphNode SVG filter semantics - no inputs, no linear

Fields

node: FilterOpGraphNode

SVGFESourceAlpha

Filter that copies the SourceAlpha image into the specified subregion, This is intentionally the only way to get SourceGraphic into the graph, as the filter region must be applied before it is used. parameters: FilterOpGraphNode SVG filter semantics - no inputs, no linear

Fields

node: FilterOpGraphNode

SVGFEIdentity

Filter that does no transformation of the colors, used for subregion cropping only.

Fields

node: FilterOpGraphNode

SVGFEOpacity

represents CSS opacity property as a graph node like the rest of the SVGFE* filters parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations

Fields

node: FilterOpGraphNode

valuebinding: PropertyBinding<f32>

value: f32

SVGFEToAlpha

convert a color image to an alpha channel - internal use; generated by SVGFilterInstance::GetOrCreateSourceAlphaIndex().

Fields

node: FilterOpGraphNode

SVGFEBlendDarken

combine 2 images with SVG_FEBLEND_MODE_DARKEN parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feBlendElement

Fields

node: FilterOpGraphNode

SVGFEBlendLighten

combine 2 images with SVG_FEBLEND_MODE_LIGHTEN parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feBlendElement

Fields

node: FilterOpGraphNode

SVGFEBlendMultiply

combine 2 images with SVG_FEBLEND_MODE_MULTIPLY parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feBlendElement

Fields

node: FilterOpGraphNode

SVGFEBlendNormal

combine 2 images with SVG_FEBLEND_MODE_NORMAL parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feBlendElement

Fields

node: FilterOpGraphNode

SVGFEBlendScreen

combine 2 images with SVG_FEBLEND_MODE_SCREEN parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feBlendElement

Fields

node: FilterOpGraphNode

SVGFEBlendOverlay

combine 2 images with SVG_FEBLEND_MODE_OVERLAY parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Source: https://developer.mozilla.org/en-US/docs/Web/CSS/mix-blend-mode

Fields

node: FilterOpGraphNode

SVGFEBlendColorDodge

combine 2 images with SVG_FEBLEND_MODE_COLOR_DODGE parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Source: https://developer.mozilla.org/en-US/docs/Web/CSS/mix-blend-mode

Fields

node: FilterOpGraphNode

SVGFEBlendColorBurn

combine 2 images with SVG_FEBLEND_MODE_COLOR_BURN parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Source: https://developer.mozilla.org/en-US/docs/Web/CSS/mix-blend-mode

Fields

node: FilterOpGraphNode

SVGFEBlendHardLight

combine 2 images with SVG_FEBLEND_MODE_HARD_LIGHT parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Source: https://developer.mozilla.org/en-US/docs/Web/CSS/mix-blend-mode

Fields

node: FilterOpGraphNode

SVGFEBlendSoftLight

combine 2 images with SVG_FEBLEND_MODE_SOFT_LIGHT parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Source: https://developer.mozilla.org/en-US/docs/Web/CSS/mix-blend-mode

Fields

node: FilterOpGraphNode

SVGFEBlendDifference

combine 2 images with SVG_FEBLEND_MODE_DIFFERENCE parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Source: https://developer.mozilla.org/en-US/docs/Web/CSS/mix-blend-mode

Fields

node: FilterOpGraphNode

SVGFEBlendExclusion

combine 2 images with SVG_FEBLEND_MODE_EXCLUSION parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Source: https://developer.mozilla.org/en-US/docs/Web/CSS/mix-blend-mode

Fields

node: FilterOpGraphNode

SVGFEBlendHue

combine 2 images with SVG_FEBLEND_MODE_HUE parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Source: https://developer.mozilla.org/en-US/docs/Web/CSS/mix-blend-mode

Fields

node: FilterOpGraphNode

SVGFEBlendSaturation

combine 2 images with SVG_FEBLEND_MODE_SATURATION parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Source: https://developer.mozilla.org/en-US/docs/Web/CSS/mix-blend-mode

Fields

node: FilterOpGraphNode

SVGFEBlendColor

combine 2 images with SVG_FEBLEND_MODE_COLOR parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Source: https://developer.mozilla.org/en-US/docs/Web/CSS/mix-blend-mode

Fields

node: FilterOpGraphNode

SVGFEBlendLuminosity

combine 2 images with SVG_FEBLEND_MODE_LUMINOSITY parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Source: https://developer.mozilla.org/en-US/docs/Web/CSS/mix-blend-mode

Fields

node: FilterOpGraphNode

SVGFEColorMatrix

transform colors of image through 5x4 color matrix (transposed for efficiency) parameters: FilterOpGraphNode, matrix[5][4] SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feColorMatrixElement

Fields

node: FilterOpGraphNode

values: [f32; 20]

SVGFEComponentTransfer

transform colors of image through configurable gradients with component swizzle parameters: FilterOpGraphNode, FilterData SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feComponentTransferElement

Fields

node: FilterOpGraphNode

SVGFECompositeArithmetic

composite 2 images with chosen composite mode with parameters for that mode parameters: FilterOpGraphNode, k1, k2, k3, k4 SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feCompositeElement

Fields

node: FilterOpGraphNode

k1: f32

k2: f32

k3: f32

k4: f32

SVGFECompositeATop

composite 2 images with chosen composite mode with parameters for that mode parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feCompositeElement

Fields

node: FilterOpGraphNode

SVGFECompositeIn

composite 2 images with chosen composite mode with parameters for that mode parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feCompositeElement

Fields

node: FilterOpGraphNode

SVGFECompositeLighter

composite 2 images with chosen composite mode with parameters for that mode parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Docs: https://developer.mozilla.org/en-US/docs/Web/SVG/Element/feComposite

Fields

node: FilterOpGraphNode

SVGFECompositeOut

composite 2 images with chosen composite mode with parameters for that mode parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feCompositeElement

Fields

node: FilterOpGraphNode

SVGFECompositeOver

composite 2 images with chosen composite mode with parameters for that mode parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feCompositeElement

Fields

node: FilterOpGraphNode

SVGFECompositeXOR

composite 2 images with chosen composite mode with parameters for that mode parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feCompositeElement

Fields

node: FilterOpGraphNode

SVGFEConvolveMatrixEdgeModeDuplicate

transform image through convolution matrix of up to 25 values (spec allows more but for performance reasons we do not) parameters: FilterOpGraphNode, orderX, orderY, kernelValues[25], divisor, bias, targetX, targetY, kernelUnitLengthX, kernelUnitLengthY, preserveAlpha SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feConvolveMatrixElement

Fields

node: FilterOpGraphNode

order_x: i32

order_y: i32

kernel: [f32; 25]

divisor: f32

bias: f32

target_x: i32

target_y: i32

kernel_unit_length_x: f32

kernel_unit_length_y: f32

preserve_alpha: i32

SVGFEConvolveMatrixEdgeModeNone

transform image through convolution matrix of up to 25 values (spec allows more but for performance reasons we do not) parameters: FilterOpGraphNode, orderX, orderY, kernelValues[25], divisor, bias, targetX, targetY, kernelUnitLengthX, kernelUnitLengthY, preserveAlpha SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feConvolveMatrixElement

Fields

node: FilterOpGraphNode

order_x: i32

order_y: i32

kernel: [f32; 25]

divisor: f32

bias: f32

target_x: i32

target_y: i32

kernel_unit_length_x: f32

kernel_unit_length_y: f32

preserve_alpha: i32

SVGFEConvolveMatrixEdgeModeWrap

transform image through convolution matrix of up to 25 values (spec allows more but for performance reasons we do not) parameters: FilterOpGraphNode, orderX, orderY, kernelValues[25], divisor, bias, targetX, targetY, kernelUnitLengthX, kernelUnitLengthY, preserveAlpha SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#feConvolveMatrixElement

Fields

node: FilterOpGraphNode

order_x: i32

order_y: i32

kernel: [f32; 25]

divisor: f32

bias: f32

target_x: i32

target_y: i32

kernel_unit_length_x: f32

kernel_unit_length_y: f32

preserve_alpha: i32

SVGFEDiffuseLightingDistant

calculate lighting based on heightmap image with provided values for a distant light source with specified direction parameters: FilterOpGraphNode, surfaceScale, diffuseConstant, kernelUnitLengthX, kernelUnitLengthY, azimuth, elevation SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEDiffuseLightingElement https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEDistantLightElement

Fields

node: FilterOpGraphNode

surface_scale: f32

diffuse_constant: f32

kernel_unit_length_x: f32

kernel_unit_length_y: f32

azimuth: f32

elevation: f32

SVGFEDiffuseLightingPoint

calculate lighting based on heightmap image with provided values for a point light source at specified location parameters: FilterOpGraphNode, surfaceScale, diffuseConstant, kernelUnitLengthX, kernelUnitLengthY, x, y, z SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEDiffuseLightingElement https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEPointLightElement

Fields

node: FilterOpGraphNode

surface_scale: f32

diffuse_constant: f32

kernel_unit_length_x: f32

kernel_unit_length_y: f32

x: f32

y: f32

z: f32

SVGFEDiffuseLightingSpot

calculate lighting based on heightmap image with provided values for a spot light source at specified location pointing at specified target location with specified hotspot sharpness and cone angle parameters: FilterOpGraphNode, surfaceScale, diffuseConstant, kernelUnitLengthX, kernelUnitLengthY, x, y, z, pointsAtX, pointsAtY, pointsAtZ, specularExponent, limitingConeAngle SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEDiffuseLightingElement https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFESpotLightElement

Fields

node: FilterOpGraphNode

surface_scale: f32

diffuse_constant: f32

kernel_unit_length_x: f32

kernel_unit_length_y: f32

x: f32

y: f32

z: f32

points_at_x: f32

points_at_y: f32

points_at_z: f32

cone_exponent: f32

limiting_cone_angle: f32

SVGFEDisplacementMap

calculate a distorted version of first input image using offset values from second input image at specified intensity parameters: FilterOpGraphNode, scale, xChannelSelector, yChannelSelector SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEDisplacementMapElement

Fields

node: FilterOpGraphNode

scale: f32

x_channel_selector: u32

y_channel_selector: u32

SVGFEDropShadow

create and merge a dropshadow version of the specified image’s alpha channel with specified offset and blur radius parameters: FilterOpGraphNode, flood_color, flood_opacity, dx, dy, stdDeviationX, stdDeviationY SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEDropShadowElement

Fields

node: FilterOpGraphNode

color: ColorF

dx: f32

dy: f32

std_deviation_x: f32

std_deviation_y: f32

SVGFEFlood

synthesize a new image of specified size containing a solid color parameters: FilterOpGraphNode, color SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEFloodElement

Fields

node: FilterOpGraphNode

color: ColorF

SVGFEGaussianBlur

create a blurred version of the input image parameters: FilterOpGraphNode, stdDeviationX, stdDeviationY SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEGaussianBlurElement

Fields

node: FilterOpGraphNode

std_deviation_x: f32

std_deviation_y: f32

SVGFEImage

synthesize a new image based on a url (i.e. blob image source) parameters: FilterOpGraphNode, sampling_filter (see SamplingFilter in Types.h), transform SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEImageElement

Fields

node: FilterOpGraphNode

sampling_filter: u32

matrix: [f32; 6]

SVGFEMorphologyDilate

create a new image based on the input image with the contour stretched outward (dilate operator) parameters: FilterOpGraphNode, radiusX, radiusY SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEMorphologyElement

Fields

node: FilterOpGraphNode

radius_x: f32

radius_y: f32

SVGFEMorphologyErode

create a new image based on the input image with the contour shrunken inward (erode operator) parameters: FilterOpGraphNode, radiusX, radiusY SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEMorphologyElement

Fields

node: FilterOpGraphNode

radius_x: f32

radius_y: f32

SVGFEOffset

create a new image that is a scrolled version of the input image, this is basically a no-op as we support offset in the graph node parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEOffsetElement

Fields

node: FilterOpGraphNode

offset_x: f32

offset_y: f32

SVGFESpecularLightingDistant

calculate lighting based on heightmap image with provided values for a distant light source with specified direction parameters: FilerData, surfaceScale, specularConstant, specularExponent, kernelUnitLengthX, kernelUnitLengthY, azimuth, elevation SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFESpecularLightingElement https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEDistantLightElement

Fields

node: FilterOpGraphNode

surface_scale: f32

specular_constant: f32

specular_exponent: f32

kernel_unit_length_x: f32

kernel_unit_length_y: f32

azimuth: f32

elevation: f32

SVGFESpecularLightingPoint

calculate lighting based on heightmap image with provided values for a point light source at specified location parameters: FilterOpGraphNode, surfaceScale, specularConstant, specularExponent, kernelUnitLengthX, kernelUnitLengthY, x, y, z SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFESpecularLightingElement https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFEPointLightElement

Fields

node: FilterOpGraphNode

surface_scale: f32

specular_constant: f32

specular_exponent: f32

kernel_unit_length_x: f32

kernel_unit_length_y: f32

x: f32

y: f32

z: f32

SVGFESpecularLightingSpot

calculate lighting based on heightmap image with provided values for a spot light source at specified location pointing at specified target location with specified hotspot sharpness and cone angle parameters: FilterOpGraphNode, surfaceScale, specularConstant, specularExponent, kernelUnitLengthX, kernelUnitLengthY, x, y, z, pointsAtX, pointsAtY, pointsAtZ, specularExponent, limitingConeAngle SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFESpecularLightingElement https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFESpotLightElement

Fields

node: FilterOpGraphNode

surface_scale: f32

specular_constant: f32

specular_exponent: f32

kernel_unit_length_x: f32

kernel_unit_length_y: f32

x: f32

y: f32

z: f32

points_at_x: f32

points_at_y: f32

points_at_z: f32

cone_exponent: f32

limiting_cone_angle: f32

SVGFETile

create a new image based on the input image, repeated throughout the output rectangle parameters: FilterOpGraphNode SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFETileElement

Fields

node: FilterOpGraphNode

SVGFETurbulenceWithFractalNoiseWithNoStitching

synthesize a new image based on Fractal Noise (Perlin) with the chosen stitching mode parameters: FilterOpGraphNode, baseFrequencyX, baseFrequencyY, numOctaves, seed SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFETurbulenceElement

Fields

node: FilterOpGraphNode

base_frequency_x: f32

base_frequency_y: f32

num_octaves: u32

seed: u32

SVGFETurbulenceWithFractalNoiseWithStitching

synthesize a new image based on Fractal Noise (Perlin) with the chosen stitching mode parameters: FilterOpGraphNode, baseFrequencyX, baseFrequencyY, numOctaves, seed SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFETurbulenceElement

Fields

node: FilterOpGraphNode

base_frequency_x: f32

base_frequency_y: f32

num_octaves: u32

seed: u32

SVGFETurbulenceWithTurbulenceNoiseWithNoStitching

synthesize a new image based on Turbulence Noise (offset vectors) parameters: FilterOpGraphNode, baseFrequencyX, baseFrequencyY, numOctaves, seed SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFETurbulenceElement

Fields

node: FilterOpGraphNode

base_frequency_x: f32

base_frequency_y: f32

num_octaves: u32

seed: u32

SVGFETurbulenceWithTurbulenceNoiseWithStitching

synthesize a new image based on Turbulence Noise (offset vectors) parameters: FilterOpGraphNode, baseFrequencyX, baseFrequencyY, numOctaves, seed SVG filter semantics - selectable input(s), selectable between linear (default) and sRGB color space for calculations Spec: https://www.w3.org/TR/filter-effects-1/#InterfaceSVGFETurbulenceElement

Fields

node: FilterOpGraphNode

base_frequency_x: f32

base_frequency_y: f32

num_octaves: u32

seed: u32