webrender_api::display_item

Struct TextDisplayItem

pub struct TextDisplayItem {
    pub common: CommonItemProperties,
    pub bounds: LayoutRect,
    pub font_key: FontInstanceKey,
    pub color: ColorF,
    pub glyph_options: Option<GlyphOptions>,
    pub ref_frame_offset: LayoutVector2D,
}

Fields
common: CommonItemPropertiesbounds: LayoutRect

The area all the glyphs should be found in. Strictly speaking this isnโt necessarily needed, but layout engines should already โknowโ this, and we use it cull and size things quickly before glyph layout is done. Currently the glyphs can be outside these bounds, but that should imply they can be cut off.

font_key: FontInstanceKeycolor: ColorFglyph_options: Option<GlyphOptions>ref_frame_offset: LayoutVector2D