Enum webrender_api::ApiMsg

pub enum ApiMsg {

    GetGlyphDimensions(GlyphDimensionRequest),
    GetGlyphIndices(GlyphIndexRequest),
    CloneApi(Sender<IdNamespace>),
    CloneApiByClient(IdNamespace),
    AddDocument(DocumentId, DeviceIntSize, DocumentLayer),
    UpdateDocuments(Vec<Box<TransactionMsg>>),
    DeleteDocument(DocumentId),
    ExternalEvent(ExternalEvent),
    ClearNamespace(IdNamespace),
    MemoryPressure,
    ReportMemory(Sender<Box<MemoryReport>>),
    DebugCommand(DebugCommand),
    WakeUp,
    WakeSceneBuilder,
    FlushSceneBuilder(Sender<()>),
    ShutDown(Option<Sender<()>>),

}

Message sent by the RenderApi to the render backend thread.

Variants

GetGlyphDimensions

Gets the glyph dimensions

Tuple Fields of GetGlyphDimensions

0: GlyphDimensionRequest

GetGlyphIndices

Gets the glyph indices from a string

Tuple Fields of GetGlyphIndices

0: GlyphIndexRequest

CloneApi

Adds a new document namespace.

Tuple Fields of CloneApi

0: Sender<IdNamespace>

CloneApiByClient

Adds a new document namespace.

Tuple Fields of CloneApiByClient

0: IdNamespace

AddDocument

Adds a new document with given initial size.

Tuple Fields of AddDocument

0: DocumentId1: DeviceIntSize2: DocumentLayer

UpdateDocuments

A message targeted at a particular document.

Tuple Fields of UpdateDocuments

0: Vec<Box<TransactionMsg>>

DeleteDocument

Deletes an existing document.

Tuple Fields of DeleteDocument

0: DocumentId

ExternalEvent

An opaque handle that must be passed to the render notifier. It is used by Gecko to forward gecko-specific messages to the render thread preserving the ordering within the other messages.

Tuple Fields of ExternalEvent

0: ExternalEvent

ClearNamespace

Removes all resources associated with a namespace.

Tuple Fields of ClearNamespace

0: IdNamespace

MemoryPressure

Flush from the caches anything that isn’t necessary, to free some memory.

ReportMemory

Collects a memory report.

Tuple Fields of ReportMemory

0: Sender<Box<MemoryReport>>

DebugCommand

Change debugging options.

Tuple Fields of DebugCommand

0: DebugCommand

WakeUp

Wakes the render backend’s event loop up. Needed when an event is communicated through another channel.

WakeSceneBuilder

See RenderApi::wake_scene_builder.

FlushSceneBuilder

Block until a round-trip to the scene builder thread has completed. This ensures that any transactions (including ones deferred to the scene builder thread) have been processed.

Tuple Fields of FlushSceneBuilder

0: Sender<()>

ShutDown

Shut the WebRender instance down.

Tuple Fields of ShutDown

0: Option<Sender<()>>