Trait webrender_api::image::AsyncBlobImageRasterizer

pub trait AsyncBlobImageRasterizer: Send {
    // Required method
    fn rasterize(
        &mut self,
        requests: &[BlobImageParams],
        low_priority: bool,
    ) -> Vec<(BlobImageRequest, BlobImageResult)>;
}

A group of rasterization requests to execute synchronously on the scene builder thread.

Required Methods

fn rasterize( &mut self, requests: &[BlobImageParams], low_priority: bool, ) -> Vec<(BlobImageRequest, BlobImageResult)>

Rasterize the requests.

Gecko uses te priority hint to schedule work in a way that minimizes the risk of high priority work being blocked by (or enqued behind) low priority work.