webrender_api

Struct BlobTilePool

pub struct BlobTilePool {
    largest_size_class: usize,
    buckets: [Vec<Arc<Vec<u8>>>; 6],
}

A pool of blob tile buffers to mitigate the overhead of allocating and deallocating blob tiles.

The pool keeps a strong reference to each allocated buffers and reuses the ones with a strong count of 1.

Fields
largest_size_class: usizebuckets: [Vec<Arc<Vec<u8>>>; 6]
Implementations

impl BlobTilePool

pub fn new() -> Self

pub fn get_buffer(&mut self, requested_size: usize) -> MutableTileBuffer

Get or allocate a tile buffer of the requested size.

The returned buffer is zero-inizitalized. The length of the returned buffer is equal to the requested size, however the buffer may be allocated with a larger capacity to conform to the pool’s corresponding bucket tile size.

fn bucket_and_size(&self, size: usize) -> (usize, usize)

pub fn cleanup(&mut self)

Go over all allocated tile buffers. For each bucket, deallocate some buffers until the number of unused buffer is more than half of the buffers for that bucket.

In practice, if called regularly, this gradually lets go of blob tiles when they are not used.