Struct webrender_api::BuiltDisplayListIter

pub struct BuiltDisplayListIter<'a> {
    list: &'a BuiltDisplayList,
    data: &'a [u8],
    cache: Option<&'a DisplayItemCache>,
    pending_items: Iter<'a, CachedDisplayItem>,
    cur_cached_item: Option<&'a CachedDisplayItem>,
    cur_item: DisplayItem,
    cur_stops: ItemRange<'a, GradientStop>,
    cur_glyphs: ItemRange<'a, GlyphInstance>,
    cur_filters: ItemRange<'a, FilterOp>,
    cur_filter_data: Vec<TempFilterData<'a>>,
    cur_filter_primitives: ItemRange<'a, FilterPrimitive>,
    cur_clip_chain_items: ItemRange<'a, ClipId>,
    cur_complex_clip: ItemRange<'a, ComplexClipRegion>,
    peeking: Peek,
    debug_stats: DebugStats,
}

Fields
list: &'a BuiltDisplayListdata: &'a [u8]cache: Option<&'a DisplayItemCache>pending_items: Iter<'a, CachedDisplayItem>cur_cached_item: Option<&'a CachedDisplayItem>cur_item: DisplayItemcur_stops: ItemRange<'a, GradientStop>cur_glyphs: ItemRange<'a, GlyphInstance>cur_filters: ItemRange<'a, FilterOp>cur_filter_data: Vec<TempFilterData<'a>>cur_filter_primitives: ItemRange<'a, FilterPrimitive>cur_clip_chain_items: ItemRange<'a, ClipId>cur_complex_clip: ItemRange<'a, ComplexClipRegion>peeking: Peekdebug_stats: DebugStats

Should just be initialized but never populated in release builds

Implementations

impl<'a> BuiltDisplayListIter<'a>

pub fn new(
list: &'a BuiltDisplayList,
data: &'a [u8],
cache: Option<&'a DisplayItemCache>
) -> Self

pub fn sub_iter(&self) -> Self

pub fn display_list(&self) -> &'a BuiltDisplayList

pub fn current_item(&self) -> &DisplayItem

fn cached_item_range_or<T>(&self, data: ItemRange<'a, T>) -> ItemRange<'a, T>

pub fn glyphs(&self) -> ItemRange<GlyphInstance>

pub fn gradient_stops(&self) -> ItemRange<GradientStop>

fn advance_pending_items(&mut self) -> bool

pub fn next<'b>(&'b mut self) -> Option<DisplayItemRef<'a, 'b>>

pub fn next_raw<'b>(&'b mut self) -> Option<DisplayItemRef<'a, 'b>>

Gets the next display item, even if it's a dummy. Also doesn't handle peeking and may leave irrelevant ranges live (so a Clip may have GradientStops if for some reason you ask).

pub fn as_ref<'b>(&'b self) -> DisplayItemRef<'a, 'b>

pub fn skip_current_stacking_context(&mut self)

pub fn current_stacking_context_empty(&mut self) -> bool

pub fn peek<'b>(&'b mut self) -> Option<DisplayItemRef<'a, 'b>>

pub fn debug_stats(&mut self) -> Vec<(&'static str, ItemStats)>

Get the debug stats for what this iterator has deserialized. Should always be empty in release builds.

pub fn merge_debug_stats_from(&mut self, other: &mut Self)

Adds the debug stats from another to our own, assuming we are a sub-iter of the other (so we can ignore where they were in the traversal).

fn log_item_stats(&mut self)