Struct webrender_api::HitTestItem

pub struct HitTestItem {
    pub pipeline: PipelineId,
    pub tag: ItemTag,
    pub point_in_viewport: LayoutPoint,
    pub point_relative_to_item: LayoutPoint,
}

Fields
pipeline: PipelineId

The pipeline that the display item that was hit belongs to.

tag: ItemTag

The tag of the hit display item.

point_in_viewport: LayoutPoint

The hit point in the coordinate space of the "viewport" of the display item. The viewport is the scroll node formed by the root reference frame of the display item's pipeline.

point_relative_to_item: LayoutPoint

The coordinates of the original hit test point relative to the origin of this item. This is useful for calculating things like text offsets in the client.