Struct webrender_api::RenderApi

pub struct RenderApi { /* fields omitted */ }

Methods

impl RenderApi

pub fn get_namespace_id(&self) -> IdNamespace

pub fn clone_sender(&self) -> RenderApiSender

pub fn add_document(
&self,
initial_size: DeviceUintSize,
layer: DocumentLayer
) -> DocumentId

pub fn delete_document(&self, document_id: DocumentId)

pub fn generate_font_key(&self) -> FontKey

pub fn generate_font_instance_key(&self) -> FontInstanceKey

pub fn get_glyph_dimensions(
&self,
font: FontInstanceKey,
glyph_indices: Vec<GlyphIndex>
) -> Vec<Option<GlyphDimensions>>

Gets the dimensions for the supplied glyph keys

Note: Internally, the internal texture cache doesn't store 'empty' textures (height or width = 0) This means that glyph dimensions e.g. for spaces (' ') will mostly be None.

pub fn get_glyph_indices(
&self,
font_key: FontKey,
text: &str
) -> Vec<Option<u32>>

Gets the glyph indices for the supplied string. These can be used to construct GlyphKeys.

pub fn generate_image_key(&self) -> ImageKey

Creates an ImageKey.

pub fn update_resources(&self, resources: Vec<ResourceUpdate>)

Add/remove/update resources such as images and fonts.

pub fn send_external_event(&self, evt: ExternalEvent)

pub fn notify_memory_pressure(&self)

pub fn shut_down(&self)

pub fn generate_property_binding_key<T: Copy>(&self) -> PropertyBindingKey<T>

Create a new unique key that can be used for animated property bindings.

pub fn send_transaction(
&self,
document_id: DocumentId,
transaction: Transaction
)

pub fn hit_test(
&self,
document_id: DocumentId,
pipeline_id: Option<PipelineId>,
point: WorldPoint,
flags: HitTestFlags
) -> HitTestResult

Does a hit test on display items in the specified document, at the given point. If a pipeline_id is specified, it is used to further restrict the hit results so that only items inside that pipeline are matched. If the HitTestFlags argument contains the FIND_ALL flag, then the vector of hit results will contain all display items that match, ordered from front to back.

pub fn set_window_parameters(
&self,
document_id: DocumentId,
window_size: DeviceUintSize,
inner_rect: DeviceUintRect,
device_pixel_ratio: f32
)

pub fn get_scroll_node_state(
&self,
document_id: DocumentId
) -> Vec<ScrollNodeState>

pub fn wake_scene_builder(&self)

pub fn flush_scene_builder(&self)

Block until a round-trip to the scene builder thread has completed. This ensures that any transactions (including ones deferred to the scene builder thread) have been processed.

pub fn save_capture(&self, path: PathBuf, bits: CaptureBits)

Save a capture of the current frame state for debugging.

pub fn load_capture(&self, path: PathBuf) -> Vec<CapturedDocument>

Load a capture of the current frame state for debugging.

pub fn send_debug_cmd(&self, cmd: DebugCommand)