Trait webrender_api::ApiHitTester

pub trait ApiHitTester: Send + Sync {
    fn hit_test(
        &self, 
        pipeline_id: Option<PipelineId>, 
        point: WorldPoint, 
        flags: HitTestFlags
    ) -> HitTestResult;
}

An object that can perform hit-testing without doing synchronous queries to the RenderBackendThread.

Required methods

fn hit_test(
&self,
pipeline_id: Option<PipelineId>,
point: WorldPoint,
flags: HitTestFlags
) -> HitTestResult

Does a hit test on display items in the specified document, at the given point. If a pipeline_id is specified, it is used to further restrict the hit results so that only items inside that pipeline are matched. If the HitTestFlags argument contains the FIND_ALL flag, then the vector of hit results will contain all display items that match, ordered from front to back.