Trait webrender_api::OutputImageHandler

pub trait OutputImageHandler {
    fn lock(
        &mut self, 
        pipeline_id: PipelineId
    ) -> Option<(u32, FramebufferIntSize)>;

    fn unlock(&mut self, pipeline_id: PipelineId);
}

Allows callers to receive a texture with the contents of a specific pipeline copied to it.

Required methods

fn lock(&mut self, pipeline_id: PipelineId) -> Option<(u32, FramebufferIntSize)>

Return the native texture handle and the size of the texture.

fn unlock(&mut self, pipeline_id: PipelineId)

Unlock will only be called if the lock() call succeeds, when WR has issued the GL commands to copy the output to the texture handle.