Type Alias webrender_api::units::ImageDirtyRect

pub type ImageDirtyRect = DirtyRect<i32, DevicePixel>;

Aliased Type

enum ImageDirtyRect {
    All,
    Partial(Box2D<i32, DevicePixel>),
}

Variants

All

Everything is Dirty, equivalent to Partial(image_bounds)

Partial(Box2D<i32, DevicePixel>)

Some specific amount is dirty

Implementations

impl<T, U> DirtyRect<T, U>where T: Copy + Clone + PartialOrd + PartialEq + Add<T, Output = T> + Sub<T, Output = T> + Zero,

pub fn empty() -> Self

Creates an empty DirtyRect (indicating nothing is invalid)

pub fn is_empty(&self) -> bool

Returns whether the dirty rect is empty

pub fn replace_with_empty(&mut self) -> Self

Replaces self with the empty rect and returns the old value.

pub fn map<F>(self, func: F) -> Selfwhere F: FnOnce(Box2D<T, U>) -> Box2D<T, U>,

Maps over the contents of Partial.

pub fn union(&self, other: &Self) -> Self

Unions the dirty rects.

pub fn intersection(&self, other: &Self) -> Self

Intersects the dirty rects.

pub fn to_subrect_of(&self, rect: &Box2D<T, U>) -> Box2D<T, U>

Converts the dirty rect into a subrect of the given one via intersection.