Module webrender_api::api

Structs

AddBlobImage

AddFontInstance

AddImage

CaptureBits
Bit flags for WR stages to store in a capture.

CapturedDocument
Information about a loaded capture of each document that is returned by RenderBackend.

ClearCache
Mask for clearing caches in debug commands.

DebugFlags

DocumentId

DocumentTransaction

DynamicProperties
When using generate_frame(), a list of PropertyValue structures can optionally be supplied to provide the current value of any animated properties.

Epoch

ExternalEvent
An opaque pointer-sized value.

HitTestFlags

HitTestItem

HitTestResult

IdNamespace

InternerSubReport

InterningMemoryReport
Memory report for interning-related data structures. cbindgen:derive-eq=false

MemoryReport
Collection of heap sizes, in bytes. cbindgen:derive-eq=false

NotificationRequest

PipelineId
From the point of view of WR, PipelineId is completely opaque and generic as long as it's clonable, serializable, comparable, and hashable.

PropertyBindingId

PropertyBindingKey
A unique key that is used for connecting animated property values to bindings in the display list.

PropertyValue
The current value of an animated property. This is supplied by the calling code.

RenderApi

RenderApiSender

ResourceId

ScrollNodeState

Transaction
A Transaction is a group of commands to apply atomically to a document.

TransactionMsg
Represents a transaction in the format sent through the channel.

UpdateBlobImage

UpdateImage

ZoomFactor
Represents a zoom factor.

Enums

AddFont

ApiMsg

Checkpoint

ClipIntern

DebugCommand

FilterDataIntern

FrameMsg

PrimitiveKeyKind
Information specific to a primitive type that uniquely identifies a primitive template by key.

PropertyBinding
A binding property can either be a specific value (the normal, non-animated case) or point to a binding location to fetch the current value from. Note that Binding has also a non-animated value, the value is used for the case where the animation is still in-delay phase (i.e. the animation doesn't produce any animation values).

ResourceUpdate

SceneMsg

ScrollClamping

ScrollLocation

Constants

_DERIVE_peek_poke_Peek_FOR_DocumentId

_DERIVE_peek_poke_Peek_FOR_IdNamespace

_DERIVE_peek_poke_Peek_FOR_PipelineId

_DERIVE_peek_poke_Peek_FOR_PropertyBinding

_DERIVE_peek_poke_Peek_FOR_PropertyBindingId

_DERIVE_peek_poke_Peek_FOR_PropertyBindingKey

_DERIVE_peek_poke_Poke_FOR_DocumentId

_DERIVE_peek_poke_Poke_FOR_IdNamespace

_DERIVE_peek_poke_Poke_FOR_PipelineId

_DERIVE_peek_poke_Poke_FOR_PropertyBinding

_DERIVE_peek_poke_Poke_FOR_PropertyBindingId

_DERIVE_peek_poke_Poke_FOR_PropertyBindingKey

_IMPL_DESERIALIZE_FOR_AddBlobImage

_IMPL_DESERIALIZE_FOR_AddFont

_IMPL_DESERIALIZE_FOR_AddFontInstance

_IMPL_DESERIALIZE_FOR_AddImage

_IMPL_DESERIALIZE_FOR_ApiMsg

_IMPL_DESERIALIZE_FOR_CaptureBits

_IMPL_DESERIALIZE_FOR_CapturedDocument

_IMPL_DESERIALIZE_FOR_Checkpoint

_IMPL_DESERIALIZE_FOR_ClearCache

_IMPL_DESERIALIZE_FOR_ClipIntern

_IMPL_DESERIALIZE_FOR_DebugCommand

_IMPL_DESERIALIZE_FOR_DebugFlags

_IMPL_DESERIALIZE_FOR_DocumentId

_IMPL_DESERIALIZE_FOR_DynamicProperties

_IMPL_DESERIALIZE_FOR_Epoch

_IMPL_DESERIALIZE_FOR_ExternalEvent

_IMPL_DESERIALIZE_FOR_FilterDataIntern

_IMPL_DESERIALIZE_FOR_FrameMsg

_IMPL_DESERIALIZE_FOR_HitTestFlags

_IMPL_DESERIALIZE_FOR_HitTestItem

_IMPL_DESERIALIZE_FOR_HitTestResult

_IMPL_DESERIALIZE_FOR_IdNamespace

_IMPL_DESERIALIZE_FOR_InternerSubReport

_IMPL_DESERIALIZE_FOR_InterningMemoryReport

_IMPL_DESERIALIZE_FOR_MemoryReport

_IMPL_DESERIALIZE_FOR_PipelineId

_IMPL_DESERIALIZE_FOR_PrimitiveKeyKind

_IMPL_DESERIALIZE_FOR_PropertyBinding

_IMPL_DESERIALIZE_FOR_PropertyBindingId

_IMPL_DESERIALIZE_FOR_PropertyBindingKey

_IMPL_DESERIALIZE_FOR_PropertyValue

_IMPL_DESERIALIZE_FOR_RenderApiSender

_IMPL_DESERIALIZE_FOR_ResourceId

_IMPL_DESERIALIZE_FOR_ResourceUpdate

_IMPL_DESERIALIZE_FOR_SceneMsg

_IMPL_DESERIALIZE_FOR_ScrollClamping

_IMPL_DESERIALIZE_FOR_ScrollLocation

_IMPL_DESERIALIZE_FOR_ScrollNodeState

_IMPL_DESERIALIZE_FOR_TransactionMsg

_IMPL_DESERIALIZE_FOR_UpdateBlobImage

_IMPL_DESERIALIZE_FOR_UpdateImage

_IMPL_DESERIALIZE_FOR_ZoomFactor

_IMPL_SERIALIZE_FOR_AddBlobImage

_IMPL_SERIALIZE_FOR_AddFont

_IMPL_SERIALIZE_FOR_AddFontInstance

_IMPL_SERIALIZE_FOR_AddImage

_IMPL_SERIALIZE_FOR_ApiMsg

_IMPL_SERIALIZE_FOR_CaptureBits

_IMPL_SERIALIZE_FOR_CapturedDocument

_IMPL_SERIALIZE_FOR_Checkpoint

_IMPL_SERIALIZE_FOR_ClearCache

_IMPL_SERIALIZE_FOR_ClipIntern

_IMPL_SERIALIZE_FOR_DebugCommand

_IMPL_SERIALIZE_FOR_DebugFlags

_IMPL_SERIALIZE_FOR_DocumentId

_IMPL_SERIALIZE_FOR_DynamicProperties

_IMPL_SERIALIZE_FOR_Epoch

_IMPL_SERIALIZE_FOR_ExternalEvent

_IMPL_SERIALIZE_FOR_FilterDataIntern

_IMPL_SERIALIZE_FOR_FrameMsg

_IMPL_SERIALIZE_FOR_HitTestFlags

_IMPL_SERIALIZE_FOR_HitTestItem

_IMPL_SERIALIZE_FOR_HitTestResult

_IMPL_SERIALIZE_FOR_IdNamespace

_IMPL_SERIALIZE_FOR_InternerSubReport

_IMPL_SERIALIZE_FOR_InterningMemoryReport

_IMPL_SERIALIZE_FOR_MemoryReport

_IMPL_SERIALIZE_FOR_PipelineId

_IMPL_SERIALIZE_FOR_PrimitiveKeyKind

_IMPL_SERIALIZE_FOR_PropertyBinding

_IMPL_SERIALIZE_FOR_PropertyBindingId

_IMPL_SERIALIZE_FOR_PropertyBindingKey

_IMPL_SERIALIZE_FOR_PropertyValue

_IMPL_SERIALIZE_FOR_RenderApiSender

_IMPL_SERIALIZE_FOR_ResourceId

_IMPL_SERIALIZE_FOR_ResourceUpdate

_IMPL_SERIALIZE_FOR_SceneMsg

_IMPL_SERIALIZE_FOR_ScrollClamping

_IMPL_SERIALIZE_FOR_ScrollLocation

_IMPL_SERIALIZE_FOR_ScrollNodeState

_IMPL_SERIALIZE_FOR_TransactionMsg

_IMPL_SERIALIZE_FOR_UpdateBlobImage

_IMPL_SERIALIZE_FOR_UpdateImage

_IMPL_SERIALIZE_FOR_ZoomFactor

Traits

NotificationHandler

RenderNotifier

Type Definitions

DocumentLayer
Documents are rendered in the ascending order of their associated layer values.

PipelineSourceId
This type carries no valuable semantics for WR. However, it reflects the fact that clients (Servo) may generate pipelines by different semi-independent sources. These pipelines still belong to the same IdNamespace and the same DocumentId. Having this extra Id field enables them to generate PipelineId without collision.

TileSize

VoidPtrToSizeFn
A C function that takes a pointer to a heap allocation and returns its size.