Module webrender_api::display_item

Structs

BackdropFilterDisplayItem
Renders a filtered region of its backdrop

BlendPrimitive

BlurPrimitive

BorderDisplayItem

BorderRadius

BorderSide

BoxShadowDisplayItem

ClearRectangleDisplayItem
Clears all colors from the area, making it possible to cut holes in the window. (useful for things like the macos frosted-glass effect).

ClipChainId

ClipChainItem
Just an abstraction for bundling up a bunch of clips into a "super clip".

ClipDisplayItem

ColorMatrixPrimitive
cbindgen:derive-eq=false

CommonItemProperties
A grouping of fields a lot of display items need, just to avoid repeating these over and over in this file.

ComplexClipRegion

ComponentTransferPrimitive

CompositePrimitive

ConicGradient

ConicGradientDisplayItem

DropShadowPrimitive

ExternalScrollId
An external identifier that uniquely identifies a scroll frame independent of its ClipId, which may change from frame to frame. This should be unique within a pipeline. WebRender makes no attempt to ensure uniqueness. The zero value is reserved for use by the root scroll node of every pipeline, which always has an external id.

FilterData

FilterPrimitive
SVG Filter Primitive. See: https://github.com/eqrion/cbindgen/issues/9 cbindgen:derive-eq=false

FloodPrimitive

Gradient

GradientDisplayItem
The area

GradientStop

HitTestDisplayItem
A minimal hit-testable item for the parent browser's convenience, and is slimmer than a RectangleDisplayItem (no color). The existence of this as a distinct item also makes it easier to inspect/debug display items.

IdentityPrimitive

IframeDisplayItem

ImageDisplayItem
This describes an image that fills the specified area. It stretches or shrinks the image as necessary. While RepeatingImageDisplayItem could otherwise provide a superset of the functionality, it has been problematic inferring the desired repetition properties when snapping changes the size of the primitive.

ImageMask

ImageMaskClipDisplayItem

LineDisplayItem

NinePatchBorder

NormalBorder

OffsetPrimitive

OpacityPrimitive

PrimitiveFlags

PushShadowDisplayItem

PushStackingContextDisplayItem

RadialGradient

RadialGradientDisplayItem

RectClipDisplayItem

RectangleDisplayItem
A solid or an animating color to draw (may not actually be a rectangle due to complex clips)

ReferenceFrame

ReferenceFrameDisplayListItem

RepeatingImageDisplayItem
This describes a background-image and its tiling. It repeats in a grid to fill the specified area.

RoundedRectClipDisplayItem

ScrollFrameDisplayItem

Shadow

SpaceAndClipInfo
Per-primitive information about the nodes in the clip tree and the spatial tree that the primitive belongs to.

SpatialId
A reference to a spatial node defining item positioning.

StackingContext

StackingContextFlags

StickyFrameDisplayItem

StickyOffsetBounds
The minimum and maximum allowable offset for a sticky frame in a single dimension.

TextDisplayItem

YuvImageDisplayItem

Enums

AlphaType

BorderDetails

BorderRadiusKind

BorderStyle

BoxShadowClipMode

ClipId
A reference to a clipping node defining how an item is clipped.

ClipMode

ColorRange

ColorSpace

ComponentTransferFuncType

CompositeOperator
Available composite operoations for the composite filter primitive

DebugDisplayItem
This is a "complete" version of the DisplayItem, with all implicit trailing arrays included, for debug serialization (captures).

DisplayItem

ExtendMode

FilterOp
CSS filter.

FilterPrimitiveInput
An input to a SVG filter primitive.

FilterPrimitiveKind
See: https://github.com/eqrion/cbindgen/issues/9 cbindgen:derive-eq=false

ImageRendering

LineOrientation

LineStyle

MixBlendMode

NinePatchBorderSource

RasterSpace
Configure whether the contents of a stacking context should be rasterized in local space or screen space. Local space rasterized pictures are typically used when we want to cache the output, and performance is important. Note that this is a performance hint only, which WR may choose to ignore.

ReferenceFrameKind

RepeatMode

ScrollSensitivity

TransformStyle

YuvColorSpace

YuvData

YuvFormat

Constants

ROOT_CLIP_ID

ROOT_REFERENCE_FRAME_SPATIAL_ID

ROOT_SCROLL_NODE_SPATIAL_ID

_DERIVE_peek_poke_Peek_FOR_AlphaType

_DERIVE_peek_poke_Peek_FOR_BackdropFilterDisplayItem

_DERIVE_peek_poke_Peek_FOR_BlendPrimitive

_DERIVE_peek_poke_Peek_FOR_BlurPrimitive

_DERIVE_peek_poke_Peek_FOR_BorderDetails

_DERIVE_peek_poke_Peek_FOR_BorderDisplayItem

_DERIVE_peek_poke_Peek_FOR_BorderRadius

_DERIVE_peek_poke_Peek_FOR_BorderRadiusKind

_DERIVE_peek_poke_Peek_FOR_BorderSide

_DERIVE_peek_poke_Peek_FOR_BorderStyle

_DERIVE_peek_poke_Peek_FOR_BoxShadowClipMode

_DERIVE_peek_poke_Peek_FOR_BoxShadowDisplayItem

_DERIVE_peek_poke_Peek_FOR_ClearRectangleDisplayItem

_DERIVE_peek_poke_Peek_FOR_ClipChainId

_DERIVE_peek_poke_Peek_FOR_ClipChainItem

_DERIVE_peek_poke_Peek_FOR_ClipDisplayItem

_DERIVE_peek_poke_Peek_FOR_ClipId

_DERIVE_peek_poke_Peek_FOR_ClipMode

_DERIVE_peek_poke_Peek_FOR_ColorMatrixPrimitive

_DERIVE_peek_poke_Peek_FOR_ColorRange

_DERIVE_peek_poke_Peek_FOR_ColorSpace

_DERIVE_peek_poke_Peek_FOR_CommonItemProperties

_DERIVE_peek_poke_Peek_FOR_ComplexClipRegion

_DERIVE_peek_poke_Peek_FOR_ComponentTransferFuncType

_DERIVE_peek_poke_Peek_FOR_ComponentTransferPrimitive

_DERIVE_peek_poke_Peek_FOR_CompositeOperator

_DERIVE_peek_poke_Peek_FOR_CompositePrimitive

_DERIVE_peek_poke_Peek_FOR_ConicGradient

_DERIVE_peek_poke_Peek_FOR_ConicGradientDisplayItem

_DERIVE_peek_poke_Peek_FOR_DisplayItem

_DERIVE_peek_poke_Peek_FOR_DropShadowPrimitive

_DERIVE_peek_poke_Peek_FOR_ExtendMode

_DERIVE_peek_poke_Peek_FOR_ExternalScrollId

_DERIVE_peek_poke_Peek_FOR_FilterOp

_DERIVE_peek_poke_Peek_FOR_FilterPrimitive

_DERIVE_peek_poke_Peek_FOR_FilterPrimitiveInput

_DERIVE_peek_poke_Peek_FOR_FilterPrimitiveKind

_DERIVE_peek_poke_Peek_FOR_FloodPrimitive

_DERIVE_peek_poke_Peek_FOR_Gradient

_DERIVE_peek_poke_Peek_FOR_GradientDisplayItem

_DERIVE_peek_poke_Peek_FOR_GradientStop

_DERIVE_peek_poke_Peek_FOR_HitTestDisplayItem

_DERIVE_peek_poke_Peek_FOR_IdentityPrimitive

_DERIVE_peek_poke_Peek_FOR_IframeDisplayItem

_DERIVE_peek_poke_Peek_FOR_ImageDisplayItem

_DERIVE_peek_poke_Peek_FOR_ImageMask

_DERIVE_peek_poke_Peek_FOR_ImageMaskClipDisplayItem

_DERIVE_peek_poke_Peek_FOR_ImageRendering

_DERIVE_peek_poke_Peek_FOR_LineDisplayItem

_DERIVE_peek_poke_Peek_FOR_LineOrientation

_DERIVE_peek_poke_Peek_FOR_LineStyle

_DERIVE_peek_poke_Peek_FOR_MixBlendMode

_DERIVE_peek_poke_Peek_FOR_NinePatchBorder

_DERIVE_peek_poke_Peek_FOR_NinePatchBorderSource

_DERIVE_peek_poke_Peek_FOR_NormalBorder

_DERIVE_peek_poke_Peek_FOR_OffsetPrimitive

_DERIVE_peek_poke_Peek_FOR_OpacityPrimitive

_DERIVE_peek_poke_Peek_FOR_PrimitiveFlags

_DERIVE_peek_poke_Peek_FOR_PushShadowDisplayItem

_DERIVE_peek_poke_Peek_FOR_PushStackingContextDisplayItem

_DERIVE_peek_poke_Peek_FOR_RadialGradient

_DERIVE_peek_poke_Peek_FOR_RadialGradientDisplayItem

_DERIVE_peek_poke_Peek_FOR_RasterSpace

_DERIVE_peek_poke_Peek_FOR_RectClipDisplayItem

_DERIVE_peek_poke_Peek_FOR_RectangleDisplayItem

_DERIVE_peek_poke_Peek_FOR_ReferenceFrame

_DERIVE_peek_poke_Peek_FOR_ReferenceFrameDisplayListItem

_DERIVE_peek_poke_Peek_FOR_ReferenceFrameKind

_DERIVE_peek_poke_Peek_FOR_RepeatMode

_DERIVE_peek_poke_Peek_FOR_RepeatingImageDisplayItem

_DERIVE_peek_poke_Peek_FOR_RoundedRectClipDisplayItem

_DERIVE_peek_poke_Peek_FOR_ScrollFrameDisplayItem

_DERIVE_peek_poke_Peek_FOR_ScrollSensitivity

_DERIVE_peek_poke_Peek_FOR_Shadow

_DERIVE_peek_poke_Peek_FOR_SpaceAndClipInfo

_DERIVE_peek_poke_Peek_FOR_SpatialId

_DERIVE_peek_poke_Peek_FOR_StackingContext

_DERIVE_peek_poke_Peek_FOR_StackingContextFlags

_DERIVE_peek_poke_Peek_FOR_StickyFrameDisplayItem

_DERIVE_peek_poke_Peek_FOR_StickyOffsetBounds

_DERIVE_peek_poke_Peek_FOR_TextDisplayItem

_DERIVE_peek_poke_Peek_FOR_TransformStyle

_DERIVE_peek_poke_Peek_FOR_YuvColorSpace

_DERIVE_peek_poke_Peek_FOR_YuvData

_DERIVE_peek_poke_Peek_FOR_YuvFormat

_DERIVE_peek_poke_Peek_FOR_YuvImageDisplayItem

_DERIVE_peek_poke_Poke_FOR_AlphaType

_DERIVE_peek_poke_Poke_FOR_BackdropFilterDisplayItem

_DERIVE_peek_poke_Poke_FOR_BlendPrimitive

_DERIVE_peek_poke_Poke_FOR_BlurPrimitive

_DERIVE_peek_poke_Poke_FOR_BorderDetails

_DERIVE_peek_poke_Poke_FOR_BorderDisplayItem

_DERIVE_peek_poke_Poke_FOR_BorderRadius

_DERIVE_peek_poke_Poke_FOR_BorderRadiusKind

_DERIVE_peek_poke_Poke_FOR_BorderSide

_DERIVE_peek_poke_Poke_FOR_BorderStyle

_DERIVE_peek_poke_Poke_FOR_BoxShadowClipMode

_DERIVE_peek_poke_Poke_FOR_BoxShadowDisplayItem

_DERIVE_peek_poke_Poke_FOR_ClearRectangleDisplayItem

_DERIVE_peek_poke_Poke_FOR_ClipChainId

_DERIVE_peek_poke_Poke_FOR_ClipChainItem

_DERIVE_peek_poke_Poke_FOR_ClipDisplayItem

_DERIVE_peek_poke_Poke_FOR_ClipId

_DERIVE_peek_poke_Poke_FOR_ClipMode

_DERIVE_peek_poke_Poke_FOR_ColorMatrixPrimitive

_DERIVE_peek_poke_Poke_FOR_ColorRange

_DERIVE_peek_poke_Poke_FOR_ColorSpace

_DERIVE_peek_poke_Poke_FOR_CommonItemProperties

_DERIVE_peek_poke_Poke_FOR_ComplexClipRegion

_DERIVE_peek_poke_Poke_FOR_ComponentTransferFuncType

_DERIVE_peek_poke_Poke_FOR_ComponentTransferPrimitive

_DERIVE_peek_poke_Poke_FOR_CompositeOperator

_DERIVE_peek_poke_Poke_FOR_CompositePrimitive

_DERIVE_peek_poke_Poke_FOR_ConicGradient

_DERIVE_peek_poke_Poke_FOR_ConicGradientDisplayItem

_DERIVE_peek_poke_Poke_FOR_DisplayItem

_DERIVE_peek_poke_Poke_FOR_DropShadowPrimitive

_DERIVE_peek_poke_Poke_FOR_ExtendMode

_DERIVE_peek_poke_Poke_FOR_ExternalScrollId

_DERIVE_peek_poke_Poke_FOR_FilterOp

_DERIVE_peek_poke_Poke_FOR_FilterPrimitive

_DERIVE_peek_poke_Poke_FOR_FilterPrimitiveInput

_DERIVE_peek_poke_Poke_FOR_FilterPrimitiveKind

_DERIVE_peek_poke_Poke_FOR_FloodPrimitive

_DERIVE_peek_poke_Poke_FOR_Gradient

_DERIVE_peek_poke_Poke_FOR_GradientDisplayItem

_DERIVE_peek_poke_Poke_FOR_GradientStop

_DERIVE_peek_poke_Poke_FOR_HitTestDisplayItem

_DERIVE_peek_poke_Poke_FOR_IdentityPrimitive

_DERIVE_peek_poke_Poke_FOR_IframeDisplayItem

_DERIVE_peek_poke_Poke_FOR_ImageDisplayItem

_DERIVE_peek_poke_Poke_FOR_ImageMask

_DERIVE_peek_poke_Poke_FOR_ImageMaskClipDisplayItem

_DERIVE_peek_poke_Poke_FOR_ImageRendering

_DERIVE_peek_poke_Poke_FOR_LineDisplayItem

_DERIVE_peek_poke_Poke_FOR_LineOrientation

_DERIVE_peek_poke_Poke_FOR_LineStyle

_DERIVE_peek_poke_Poke_FOR_MixBlendMode

_DERIVE_peek_poke_Poke_FOR_NinePatchBorder

_DERIVE_peek_poke_Poke_FOR_NinePatchBorderSource

_DERIVE_peek_poke_Poke_FOR_NormalBorder

_DERIVE_peek_poke_Poke_FOR_OffsetPrimitive

_DERIVE_peek_poke_Poke_FOR_OpacityPrimitive

_DERIVE_peek_poke_Poke_FOR_PrimitiveFlags

_DERIVE_peek_poke_Poke_FOR_PushShadowDisplayItem

_DERIVE_peek_poke_Poke_FOR_PushStackingContextDisplayItem

_DERIVE_peek_poke_Poke_FOR_RadialGradient

_DERIVE_peek_poke_Poke_FOR_RadialGradientDisplayItem

_DERIVE_peek_poke_Poke_FOR_RasterSpace

_DERIVE_peek_poke_Poke_FOR_RectClipDisplayItem

_DERIVE_peek_poke_Poke_FOR_RectangleDisplayItem

_DERIVE_peek_poke_Poke_FOR_ReferenceFrame

_DERIVE_peek_poke_Poke_FOR_ReferenceFrameDisplayListItem

_DERIVE_peek_poke_Poke_FOR_ReferenceFrameKind

_DERIVE_peek_poke_Poke_FOR_RepeatMode

_DERIVE_peek_poke_Poke_FOR_RepeatingImageDisplayItem

_DERIVE_peek_poke_Poke_FOR_RoundedRectClipDisplayItem

_DERIVE_peek_poke_Poke_FOR_ScrollFrameDisplayItem

_DERIVE_peek_poke_Poke_FOR_ScrollSensitivity

_DERIVE_peek_poke_Poke_FOR_Shadow

_DERIVE_peek_poke_Poke_FOR_SpaceAndClipInfo

_DERIVE_peek_poke_Poke_FOR_SpatialId

_DERIVE_peek_poke_Poke_FOR_StackingContext

_DERIVE_peek_poke_Poke_FOR_StackingContextFlags

_DERIVE_peek_poke_Poke_FOR_StickyFrameDisplayItem

_DERIVE_peek_poke_Poke_FOR_StickyOffsetBounds

_DERIVE_peek_poke_Poke_FOR_TextDisplayItem

_DERIVE_peek_poke_Poke_FOR_TransformStyle

_DERIVE_peek_poke_Poke_FOR_YuvColorSpace

_DERIVE_peek_poke_Poke_FOR_YuvData

_DERIVE_peek_poke_Poke_FOR_YuvFormat

_DERIVE_peek_poke_Poke_FOR_YuvImageDisplayItem

Functions

sanitize_func_type

sanitize_values

Type Definitions

ItemKey
An identifier used to refer to previously sent display items. Currently it refers to individual display items, but this may change later.

ItemTag
A tag that can be used to identify items during hit testing. If the tag is missing then the item doesn't take part in hit testing at all. This is composed of two numbers. In Servo, the first is an identifier while the second is used to select the cursor that should be used during mouse movement. In Gecko, the first is a scrollframe identifier, while the second is used to store various flags that APZ needs to properly process input events.