Module webrender_api::display_item

Macros

impl_default_for_enums 🔒

Structs

BackdropFilterDisplayItem

Renders a filtered region of its backdrop

BlendPrimitive

BlurPrimitive

BorderDisplayItem

BorderRadius

BorderSide

BoxShadowDisplayItem

ClearRectangleDisplayItem

Clears all colors from the area, making it possible to cut holes in the window. (useful for things like the macos frosted-glass effect).

ClipChainId

ClipChainItem

Just an abstraction for bundling up a bunch of clips into a “super clip”.

ClipId

A reference to a clipping node defining how an item is clipped.

ColorMatrixPrimitive

cbindgen:derive-eq=false

CommonItemProperties

A grouping of fields a lot of display items need, just to avoid repeating these over and over in this file.

ComplexClipRegion

ComponentTransferPrimitive

CompositePrimitive

ConicGradient

ConicGradientDisplayItem

DropShadowPrimitive

ExternalScrollId

An external identifier that uniquely identifies a scroll frame independent of its ClipId, which may change from frame to frame. This should be unique within a pipeline. WebRender makes no attempt to ensure uniqueness. The zero value is reserved for use by the root scroll node of every pipeline, which always has an external id.

FilterData

FilterPrimitive

SVG Filter Primitive. See: https://github.com/eqrion/cbindgen/issues/9 cbindgen:derive-eq=false

FloodPrimitive

Gradient

GradientDisplayItem

The area

GradientStop

HitTestDisplayItem

A minimal hit-testable item for the parent browser’s convenience, and is slimmer than a RectangleDisplayItem (no color). The existence of this as a distinct item also makes it easier to inspect/debug display items.

IdentityPrimitive

IframeDisplayItem

ImageDisplayItem

This describes an image that fills the specified area. It stretches or shrinks the image as necessary. While RepeatingImageDisplayItem could otherwise provide a superset of the functionality, it has been problematic inferring the desired repetition properties when snapping changes the size of the primitive.

ImageMask

ImageMaskClipDisplayItem

LineDisplayItem

NinePatchBorder

NormalBorder

OffsetPrimitive

OpacityPrimitive

PrimitiveFlags

PushShadowDisplayItem

PushStackingContextDisplayItem

RadialGradient

RadialGradientDisplayItem

RectClipDisplayItem

RectangleDisplayItem

A solid or an animating color to draw (may not actually be a rectangle due to complex clips)

ReferenceFrame

ReferenceFrameDescriptor

ReferenceFrameDisplayListItem

RepeatingImageDisplayItem

This describes a background-image and its tiling. It repeats in a grid to fill the specified area.

RoundedRectClipDisplayItem

ScrollFrameDescriptor

Shadow

SpaceAndClipInfo

Per-primitive information about the nodes in the clip tree and the spatial tree that the primitive belongs to.

SpatialId

A reference to a spatial node defining item positioning.

SpatialTreeItemKey

Defines a caller provided key that is unique for a given spatial node, and is stable across display lists. WR uses this to determine which spatial nodes are added / removed for a new display list. The content itself is arbitrary and opaque to WR, the only thing that matters is that it’s unique and stable between display lists.

StackingContext

StackingContextFlags

StickyFrameDescriptor

StickyOffsetBounds

The minimum and maximum allowable offset for a sticky frame in a single dimension.

TextDisplayItem

YuvImageDisplayItem

Enums

AlphaType

BorderDetails

BorderRadiusKind

BorderStyle

BoxShadowClipMode

ClipMode

ColorRange

ColorSpace

ComponentTransferFuncType

CompositeOperator

Available composite operoations for the composite filter primitive

DebugDisplayItem

This is a “complete” version of the DisplayItem, with all implicit trailing arrays included, for debug serialization (captures).

DisplayItem

ExtendMode

FillRule

FilterOp

CSS filter.

FilterPrimitiveInput

An input to a SVG filter primitive.

FilterPrimitiveKind

See: https://github.com/eqrion/cbindgen/issues/9 cbindgen:derive-eq=false

ImageRendering

LineOrientation

LineStyle

MixBlendMode

NinePatchBorderSource

RasterSpace

Configure whether the contents of a stacking context should be rasterized in local space or screen space. Local space rasterized pictures are typically used when we want to cache the output, and performance is important. Note that this is a performance hint only, which WR may choose to ignore.

ReferenceFrameKind

ReferenceTransformBinding

RepeatMode

Rotation

SpatialTreeItem

TransformStyle

YuvColorSpace

YuvData

YuvFormat

YuvRangedColorSpace

Constants

POLYGON_CLIP_VERTEX_MAX

ROOT_CLIP_ID 🔒

ROOT_REFERENCE_FRAME_SPATIAL_ID 🔒

ROOT_SCROLL_NODE_SPATIAL_ID 🔒

Functions

sanitize_func_type 🔒

sanitize_values 🔒

Type Aliases

ItemKey

An identifier used to refer to previously sent display items. Currently it refers to individual display items, but this may change later.

ItemTag

A tag that can be used to identify items during hit testing. If the tag is missing then the item doesn’t take part in hit testing at all. This is composed of two numbers. In Servo, the first is an identifier while the second is used to select the cursor that should be used during mouse movement. In Gecko, the first is a scrollframe identifier, while the second is used to store various flags that APZ needs to properly process input events.