Module webrender_api::display_list

Structs

AuxIter

BuiltDisplayList
A display list.

BuiltDisplayListDescriptor
Describes the memory layout of a display list.

BuiltDisplayListIter

DebugStats
Internal info used for more detailed analysis of serialized display lists

DisplayItemRef

DisplayListBuilder

DisplayListWithCache

ItemRange

ItemStats
Stats for an individual item

SaveState

TempFilterData

Enums

DisplayListSection
DisplayListSection determines the target buffer for the display items.

Peek

Constants

FIRST_CLIP_NODE_INDEX

FIRST_SPATIAL_NODE_INDEX

MAX_TEXT_RUN_LENGTH

Functions

skip_slice
Returns the byte-range the slice occupied.