Module webrender_api::display_list

Structs

AuxIter

BuiltDisplayList

A display list.

BuiltDisplayListDescriptor

Describes the memory layout of a display list.

BuiltDisplayListIter

DebugStats 🔒

Internal info used for more detailed analysis of serialized display lists

DisplayItemRef

DisplayListBuilder

DisplayListCapacity 🔒

DisplayListCapture 🔒

A debug (human-readable) representation of a built display list that can be used for capture and replay.

DisplayListPayload

DisplayListWithCache

ItemRange

ItemStats

Stats for an individual item

ReferenceFrameMapper

Maps from stacking context layout coordinates into reference frame relative coordinates.

ReferenceFrameState 🔒

The offset stack for a given reference frame.

SaveState

SpatialNodeInfo

A small portion of a normal spatial node that we store during DL construction to enable snapping and reference frame <-> stacking context coord mapping. In future we’ll aim to remove this and have the full spatial tree available during DL build.

TempFilterData

Enums

BuildState 🔒

DisplayListSection

DisplayListSection determines the target buffer for the display items.

GeckoDisplayListType

Peek 🔒

Constants

FIRST_CLIP_NODE_INDEX 🔒

FIRST_SPATIAL_NODE_INDEX 🔒

MAX_TEXT_RUN_LENGTH

Functions

iter_spatial_tree 🔒

skip_slice 🔒

Returns the byte-range the slice occupied.