Enum webrender_api::ApiMsg

pub enum ApiMsg {
    UpdateResources(ResourceUpdates),
    GetGlyphDimensions(FontInstanceKey, Vec<GlyphKey>, MsgSender<Vec<Option<GlyphDimensions>>>),
    GetGlyphIndices(FontKey, String, MsgSender<Vec<Option<u32>>>),
    CloneApi(MsgSender<IdNamespace>),
    AddDocument(DocumentId, DeviceUintSize, DocumentLayer),
    UpdateDocument(DocumentId, Vec<DocumentMsg>),
    DeleteDocument(DocumentId),
    ExternalEvent(ExternalEvent),
    ClearNamespace(IdNamespace),
    MemoryPressure,
    DebugCommand(DebugCommand),
    ShutDown,
}

Variants
UpdateResources(ResourceUpdates)

Add/remove/update images and fonts.

GetGlyphDimensions(FontInstanceKey, Vec<GlyphKey>, MsgSender<Vec<Option<GlyphDimensions>>>)

Gets the glyph dimensions

GetGlyphIndices(FontKey, String, MsgSender<Vec<Option<u32>>>)

Gets the glyph indices from a string

CloneApi(MsgSender<IdNamespace>)

Adds a new document namespace.

AddDocument(DocumentId, DeviceUintSize, DocumentLayer)

Adds a new document with given initial size.

UpdateDocument(DocumentId, Vec<DocumentMsg>)

A message targeted at a particular document.

DeleteDocument(DocumentId)

Deletes an existing document.

ExternalEvent(ExternalEvent)

An opaque handle that must be passed to the render notifier. It is used by Gecko to forward gecko-specific messages to the render thread preserving the ordering within the other messages.

ClearNamespace(IdNamespace)

Removes all resources associated with a namespace.

MemoryPressure

Flush from the caches anything that isn't necessary, to free some memory.

DebugCommand(DebugCommand)

Change debugging options.

ShutDown