webrender_api

Module font

Structs

FontInstanceData

Data corresponding to an instantiation of a font, with size and other options specified.

FontInstanceFlags

FontInstanceKey

FontInstanceOptions

FontInstancePlatformOptions

FontKey

FontSize

Hashable floating-point storage for font size.

FontVariation

GlyphDimensionRequest

GlyphDimensions

GlyphIndexRequest

GlyphInstance

GlyphOptions

NativeFontHandle

SyntheticItalics

Enums

FontHinting

FontLCDFilter

FontRenderMode

FontTemplate

Container for the raw data describing a font. This might be a stream of bytes corresponding to a downloaded font, or a handle to a native font from the operating system.

Type Aliases

GlyphIndex