Module webrender_api::font

Structs

FontInstanceData
Data corresponding to an instantiation of a font, with size and other options specified.

FontInstanceFlags

FontInstanceKey

FontInstanceOptions

FontInstancePlatformOptions

FontKey

FontVariation

GlyphDimensions

GlyphInstance

GlyphOptions

NativeFontHandle

SyntheticItalics

Enums

FontHinting

FontLCDFilter

FontRenderMode

FontTemplate
Container for the raw data describing a font. This might be a stream of bytes corresponding to a downloaded font, or a handle to a native font from the operating system.

Constants

_DERIVE_peek_poke_Peek_FOR_FontInstanceFlags

_DERIVE_peek_poke_Peek_FOR_FontInstanceKey

_DERIVE_peek_poke_Peek_FOR_FontRenderMode

_DERIVE_peek_poke_Peek_FOR_GlyphInstance

_DERIVE_peek_poke_Peek_FOR_GlyphOptions

_DERIVE_peek_poke_Poke_FOR_FontInstanceFlags

_DERIVE_peek_poke_Poke_FOR_FontInstanceKey

_DERIVE_peek_poke_Poke_FOR_FontRenderMode

_DERIVE_peek_poke_Poke_FOR_GlyphInstance

_DERIVE_peek_poke_Poke_FOR_GlyphOptions

_IMPL_DESERIALIZE_FOR_FontHinting

_IMPL_DESERIALIZE_FOR_FontInstanceFlags

_IMPL_DESERIALIZE_FOR_FontInstanceKey

_IMPL_DESERIALIZE_FOR_FontInstanceOptions

_IMPL_DESERIALIZE_FOR_FontInstancePlatformOptions

_IMPL_DESERIALIZE_FOR_FontKey

_IMPL_DESERIALIZE_FOR_FontLCDFilter

_IMPL_DESERIALIZE_FOR_FontRenderMode

_IMPL_DESERIALIZE_FOR_FontVariation

_IMPL_DESERIALIZE_FOR_GlyphDimensions

_IMPL_DESERIALIZE_FOR_GlyphInstance

_IMPL_DESERIALIZE_FOR_GlyphOptions

_IMPL_DESERIALIZE_FOR_NativeFontHandle

_IMPL_DESERIALIZE_FOR_SyntheticItalics

_IMPL_SERIALIZE_FOR_FontHinting

_IMPL_SERIALIZE_FOR_FontInstanceFlags

_IMPL_SERIALIZE_FOR_FontInstanceKey

_IMPL_SERIALIZE_FOR_FontInstanceOptions

_IMPL_SERIALIZE_FOR_FontInstancePlatformOptions

_IMPL_SERIALIZE_FOR_FontKey

_IMPL_SERIALIZE_FOR_FontLCDFilter

_IMPL_SERIALIZE_FOR_FontRenderMode

_IMPL_SERIALIZE_FOR_FontVariation

_IMPL_SERIALIZE_FOR_GlyphDimensions

_IMPL_SERIALIZE_FOR_GlyphInstance

_IMPL_SERIALIZE_FOR_GlyphOptions

_IMPL_SERIALIZE_FOR_NativeFontHandle

_IMPL_SERIALIZE_FOR_SyntheticItalics

Type Definitions

GlyphIndex