Module webrender_api::font

Structs

BaseFontInstance

Immutable description of a font instance requested by the user of the API.

FontInstanceData

Data corresponding to an instantiation of a font, with size and other options specified.

FontInstanceFlags

FontInstanceKey

FontInstanceOptions

FontInstancePlatformOptions

FontKey

FontSize

Hashable floating-point storage for font size.

FontVariation

GlyphDimensionRequest

GlyphDimensions

GlyphIndexRequest

GlyphInstance

GlyphOptions

NativeFontHandle

SharedFontInstanceMap

A map of font instance data accessed concurrently from multiple threads.

SyntheticItalics

Enums

FontHinting

FontLCDFilter

FontRenderMode

FontTemplate

Container for the raw data describing a font. This might be a stream of bytes corresponding to a downloaded font, or a handle to a native font from the operating system.

Constants

_DERIVE_peek_poke_Peek_FOR_FontInstanceFlags

_DERIVE_peek_poke_Peek_FOR_FontInstanceKey

_DERIVE_peek_poke_Peek_FOR_FontRenderMode

_DERIVE_peek_poke_Peek_FOR_GlyphInstance

_DERIVE_peek_poke_Peek_FOR_GlyphOptions

_DERIVE_peek_poke_Poke_FOR_FontInstanceFlags

_DERIVE_peek_poke_Poke_FOR_FontInstanceKey

_DERIVE_peek_poke_Poke_FOR_FontRenderMode

_DERIVE_peek_poke_Poke_FOR_GlyphInstance

_DERIVE_peek_poke_Poke_FOR_GlyphOptions

Type Definitions

FontInstanceMap

GlyphIndex