Struct webrender_api::Transaction

pub struct Transaction { /* fields omitted */ }

A Transaction is a group of commands to apply atomically to a document.

This mechanism ensures that: - no other message can be interleaved between two commands that need to be applied together. - no redundant work is performed if two commands in the same transaction cause the scene or the frame to be rebuilt.

Methods

impl Transaction

pub fn new() -> Self

pub fn is_empty(&self) -> bool

pub fn update_epoch(&mut self, pipeline_id: PipelineId, epoch: Epoch)

pub fn set_root_pipeline(&mut self, pipeline_id: PipelineId)

Sets the root pipeline.

Examples

let pipeline_id = PipelineId(0, 0);
let mut txn = Transaction::new();
txn.set_root_pipeline(pipeline_id);

pub fn remove_pipeline(&mut self, pipeline_id: PipelineId)

Removes data associated with a pipeline from the internal data structures. If the specified pipeline_id is for the root pipeline, the root pipeline is reset back to None.

pub fn set_display_list(
&mut self,
epoch: Epoch,
background: Option<ColorF>,
viewport_size: LayoutSize,
(pipeline_id, content_size, display_list): (PipelineId, LayoutSize, BuiltDisplayList),
preserve_frame_state: bool
)

Supplies a new frame to WebRender.

Non-blocking, it notifies a worker process which processes the display list. When it's done and a RenderNotifier has been set in webrender::Renderer, new_frame_ready() gets called.

Note: Scrolling doesn't require an own Frame.

Arguments:

document_id: Target Document ID.

epoch: The unique Frame ID, monotonically increasing.

background: The background color of this pipeline.

viewport_size: The size of the viewport for this frame.

pipeline_id: The ID of the pipeline that is supplying this display list.

content_size: The total screen space size of this display list's display items.

display_list: The root Display list used in this frame.

preserve_frame_state: If a previous frame exists which matches this pipeline id, this setting determines if frame state (such as scrolling position) should be preserved for this new display list.

pub fn update_resources(&mut self, resources: ResourceUpdates)

pub fn set_window_parameters(
&mut self,
window_size: DeviceUintSize,
inner_rect: DeviceUintRect,
device_pixel_ratio: f32
)

pub fn scroll(
&mut self,
scroll_location: ScrollLocation,
cursor: WorldPoint,
phase: ScrollEventPhase
)

Scrolls the scrolling layer under the cursor

WebRender looks for the layer closest to the user which has ScrollPolicy::Scrollable set.

pub fn scroll_node_with_id(
&mut self,
origin: LayoutPoint,
id: ScrollNodeIdType,
clamp: ScrollClamping
)

pub fn set_page_zoom(&mut self, page_zoom: ZoomFactor)

pub fn set_pinch_zoom(&mut self, pinch_zoom: ZoomFactor)

pub fn set_pan(&mut self, pan: DeviceIntPoint)

pub fn tick_scrolling_bounce_animations(&mut self)

pub fn generate_frame(&mut self)

Generate a new frame.

pub fn update_dynamic_properties(&mut self, properties: DynamicProperties)

Supply a list of animated property bindings that should be used to resolve bindings in the current display list.

pub fn enable_frame_output(&mut self, pipeline_id: PipelineId, enable: bool)

Enable copying of the output of this pipeline id to an external texture for callers to consume.