Trait webrender_api::RenderNotifier

pub trait RenderNotifier: Send {
    fn clone(&self) -> Box<dyn RenderNotifier>;

    fn wake_up(&self);

    fn new_frame_ready(
        &self, 
        _: DocumentId, 
        scrolled: bool, 
        composite_needed: bool, 
        render_time_ns: Option<u64>
    );

    fn external_event(&self, _evt: ExternalEvent) { ... }

    fn shut_down(&self) { ... }
}

A handler to integrate WebRender with the thread that contains the Renderer.

Required methods

fn clone(&self) -> Box<dyn RenderNotifier>

fn wake_up(&self)

Wake the thread containing the Renderer up (after updates have been put in the renderer’s queue).

fn new_frame_ready(
&self,
_: DocumentId,
scrolled: bool,
composite_needed: bool,
render_time_ns: Option<u64>
)

Notify the thread containing the Renderer that a new frame is ready.

Provided methods

fn external_event(&self, _evt: ExternalEvent)

A Gecko-specific notification mechanism to get some code executed on the Renderer’s thread, mostly replaced by NotificationHandler. You should probably use the latter instead.

fn shut_down(&self)

Notify the thread containing the Renderer that the render backend has been shut down.